Key to link image based answers.
Answers to some of the questions are in the form of check boxes which are scraped as images.
The name of each image provides information on the image (e.g. whether the box is checked or not).
gif_key.csv provides the key to link each image-answer to its respective question. 
It is the only dataset required to run VisaDoorScraper.R